Rename pacmod_game_control files
